requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
reportlab==3.6.12
Pillow==9.5.0
requests==2.29.0
pipreqs==0.4.13